+++ src/NoSeekStream.php
@@ -12,6 +12,7 @@ use Psr\Http\Message\StreamInterface;
 final class NoSeekStream implements StreamInterface
 {
     use StreamDecoratorTrait;
+    use NonSerializableStreamTrait;
 
     private StreamInterface $stream;